Refactor : 안 쓰는 import나 엔터 정리
