+++ src/test/java/lotto/Exception/ExceptionControllerTest.java
@@ -1,17 +1,5 @@
 package lotto.Exception;
 
-import lotto.model.BonusNumber;
-import lotto.model.Price;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import static lotto.exception.ExceptionController.*;
-import static lotto.exception.PriceException.checkPriceException;
-
 public class ExceptionControllerTest {
